Timercontroller now handling timerlist. (This class right now might be handling way too much and can be improved later on, but this is what i came up with before testday 2).
Also, the timer is only being saved locally for now, as i don't have an account to test this on yet. There's and if-case in the saveTimerToServer() method that i chose to comment away as it will signal that the timer has not been saved properly to the server. As I don't have any time to fix this now, this will do for the tests.